+++ CacheSystem/LFUCachePolicy.cpp
@@ -4,19 +4,19 @@ namespace CacheSystem
 {
 	namespace CachePolicies
 	{
-		void LFUCachePolicy::dataCreationEvent(uint64_t dataId, uint64_t dataSize, int64_t dataCreationTime)
+		void LFUCachePolicy::dataCreationEvent(uint64_t dataId, size_t dataSize, int64_t dataCreationTime)
 		{
 			CacheData* data = (CacheData*)dataId;
 			data->setUserData(new double(0));
 		}
 
-		void LFUCachePolicy::cacheHitEvent(uint64_t dataId, uint64_t dataSize, int64_t dataCreationTime)
+		void LFUCachePolicy::cacheHitEvent(uint64_t dataId, size_t dataSize, int64_t dataCreationTime)
 		{
 			CacheData* data = (CacheData*)dataId;
 			(*(double*)data->getUserData())++;
 		}
 
-		double LFUCachePolicy::getPriority(uint64_t dataId, uint64_t dataSize, int64_t dataCreationTime)
+		double LFUCachePolicy::getPriority(uint64_t dataId, size_t dataSize, int64_t dataCreationTime)
 		{
 			CacheData* data = (CacheData*)dataId;
 			return *(double*)data->getUserData();
